Fixed build errors and warnings  on xcode, Added get euler angle method to quaternion class

+++ src/uslscore/USFileStream.cpp
@@ -42,7 +42,7 @@ ZLFILE* USFileStream::GetFile () {
 }
 
 //----------------------------------------------------------------//
-u32 USFileStream::GetLength () {
+size_t USFileStream::GetLength () {
 
 	return this->mLength;
 }
